+++ packages/be-ts-db-lib/vitest.config.ts
@@ -1,10 +1,13 @@
-import { defineProject } from 'vitest/config';
+import { defineProject, mergeConfig } from 'vitest/config';
+import { sharedProjectConfig } from '../../vitest.shared.js';
 
-export default defineProject({
-  test: {
-    exclude: ['lib/**/*', 'node_modules/**/*'],
-    globalSetup: './test-util/vitest.setup.ts',
-    // So we don't hammer the database too hard and start having tests timeout
-    maxWorkers: 4
-  }
-});
+export default mergeConfig(
+  sharedProjectConfig,
+  defineProject({
+    test: {
+      globalSetup: './test-util/vitest.setup.ts',
+      // So we don't hammer the database too hard and start having tests timeout
+      maxWorkers: 4
+    }
+  })
+);