+++ packages/core-ts-lib/vitest.config.ts
@@ -1,8 +1,4 @@
-import { defineProject } from 'vitest/config';
+import { mergeConfig } from 'vitest/config';
+import { sharedProjectConfig } from '../../vitest.shared.js';
 
-export default defineProject({
-  test: {
-    exclude: ['lib/**/*', 'node_modules/**/*'],
-    testTimeout: 10000 // 10 seconds per test
-  }
-});
+export default mergeConfig(sharedProjectConfig, {});