+++ packages/be-ts-db-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [4.2.31] (2026-06-19)
+
+### 🏗️ Changed
+
+- Updated `@aneuhold/be-ts-lib` to `^3.1.17`, `@aneuhold/core-ts-db-lib` to `^5.0.10`, and `@aneuhold/core-ts-lib` to `^2.4.6`.
+
 ## 🔖 [4.2.30] (2026-06-12)
 
 ### 🏗️ Changed
@@ -439,6 +445,7 @@ Updated dependencies: now requires `@aneuhold/core-ts-db-lib@^3.0.0`, `@aneuhold
 - Updated workflow permissions to allow repository write access
 
 <!-- Link References -->
+[4.2.31]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.30...be-ts-db-lib-v4.2.31
 [4.2.30]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.29...be-ts-db-lib-v4.2.30
 [4.2.29]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.28...be-ts-db-lib-v4.2.29
 [4.2.28]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.27...be-ts-db-lib-v4.2.28